requirements.txt
yagmail==0.15.293
python-dotenv==1.0.1
